requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
captcha==0.6.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
decorator==4.4.2
Django==4.2.15
django-multi-captcha-admin==2.0.0
django-ranged-response==0.2.0
django-recaptcha==4.0.0
django-simple-captcha==0.6.0
django-taggit==6.1.0
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
jalali_core==1.0.0
jdatetime==5.0.0
moviepy==1.0.3
numpy==1.24.4
persiantools==4.2.0
pillow==10.4.0
proglog==0.1.10
pycparser==2.22
pytz==2025.1
pyvips==2.2.3
requests==2.32.3
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3